feat(rcnn-segmenter): added implementation of detectron2 for instance segmentation (#5)
This branch implements instance segmentation on images, for general real world objects, by using an RCNN pre-trained model (more specifically the X101-FPN-3X model found here). from Facebook research's Detectron2 library. The model is used on a temporary image of a living room and visualises the segmentation of the objects, supported by the pre-trained model, in the image.