feat(depth-estimator): added estimation of projection quantities for objects (#6)
The code is now able to estimate the depth of each object segmented by the RCNN model. It currently only supports "chairs" and "couches", although this can be adjusted to support a few other furniture items. The code is also able to estimate the displacement of the object along the x-axis. It assumes all objects lie on the same y-axis (i.e the same flooring). The functions calculating the projection quantities like depth and displacement have been fully tested with unit tests.