feat(auto-furniture-placement): added automatic room and furniture placement and positioning (#8)
The code is now able to create a 3D application which opens a window after pressing a key twice in the image segmentation window. This 3D application automatically chooses the appropriate supported furniture detected in the image (chair or couch is supported at the moment) and positions them in their estimated positions based on the image. A room is also generated where its centre is automatically positioned at the average of the centres of all the furniture. It is also scaled based on the largest range between the displacements of the furnitures and the depths of the furnitures. This ensures all the furniture is captured within the generated room. As well as generating the scene automatically, the user is also able to control the camera viewing the scene, with the same controls supported by the default camera driver in the Panda3D library. When the window is closed, the application ends.