feat(image-browse-input): added gui app supporting image browsing as input (#9)
The code now runs a GUI application which contains a browse button widget that lets the user browse the file system for an input image. This input image can be selected and is then processed by the `ImageProcessor` and a new Panda3D window is opened showing the generated 3D scene.